requirements_conda.txt
pytorch=1.1.0=py3.7_cuda10.0.130_cudnn7.5.1_0
torchfile=0.1.0=pypi_0
torchvision=0.3.0=py37_cu10.0.130_1
tensorboard=2.1.0=pypi_0
tensorboardx=1.9=pypi_0
numpy=1.17.2=pypi_0
pillow=5.0.0=pypi_0
opencv=3.4.2=py37h6fd60c2_1
opencv-python=3.4.5.20=pypi_0
scipy=1.0.1=pypi_0
numpy=1.17.2=pypi_0
pillow=5.0.0=pypi_0
protobuf=3.11.2=pypi_0
six=1.13.0=py37_0
torchvision=0.3.0=py37_cu10.0.130_1
tqdm=4.35.0=pypi_0
